etl/requirements.txt
apache-airflow==2.8.0
pandas==2.0.0
requests==2.31.0
boto3==1.28.0
python-dotenv==1.0.0
kafka-python==2.0.2
confluent-kafka==2.3.0
kafka-python==2.0.2
confluent-kafka==2.3.0
EOF
